bash_profile and slow NVM handling updates

+++ refresh.sh
@@ -9,9 +9,15 @@ git pull && git submodule init && git submodule update
 
 #files that where added after install script was first created and may not be
 #present in newer versions
-files="desk ctags agignore inputrc"
+files="desk ctags agignore inputrc bash_profile"
 for file in $files; do
-	[[ -L ~/.$file ]] || ln -s ~/dotfiles/$file ~/.$file  
+	if [[ ! -L ~/.$file ]]; then
+		if [[ -f ~/.$file ]]; then
+			[[ -d ~/dotfiles_old ]] || mkdir ~/dotfiles_old
+			mv ~/.$file ~/dotfiles_old
+		fi
+		[[ -L ~/.$file ]] || ln -s ~/dotfiles/$file ~/.$file  
+	fi
 done
 
 config_files="nvim"
